Omit constr in Show/Read insts of NoDup, Bag, Set
Omit constructor in Show/Read instances of NoDup, Bag, Set and Map.

This makes them consistent with other types in the Types module.  Also, this
allows for nicer output when using the "check" function.

I may revert this though...  As ommiting the constructor can sometimes be
confusing and breaks copy-pasting.  If I ever revert this, I may also use a
standard Show instance for the "Xs" type.