Improve Show instance of functions by using _
Before, print (||) would show:

    \x y -> case (x,y) of
            (False,False) -> False
            (False,True) -> True
            (True,False) -> True
            (True,True) -> True

Now, print (||) just shows:

    \x y -> case (x,y) of
            (False,False) -> False
            (_,_) -> True

As with the previous function-showing algorithm, this is done heuristically and
is unsound in general (you may get a wrong function description).  But, it
works well on most cases.

There are still a few rough edges before I call this done.  These are listed on
the `TODO.md` file.