dev version bump
This version 0.9.5 is not intended for Hackage release
and will not receive a git tag.

From now on, releases ending in odd numbers are versions in development.

This prevents someone from accidentally releasing a version v0.9.4
by checking out the source tree a few commits after v0.9.4.

During the lifetime of v0.9.5,
the code starts exactly the same as v0.9.4
and ends exactly the same as the next version (v0.9.6).