make GHC 9.8 happy replacing by tail with drop 1
I could use -Wno-x-partial but then that would yield a warning of unrecognized
warning flags on earlier versions.  Replacing with `drop 1` seems to be the
easiest option here...

Though in this case I would actually appreciate the
error that tail throws!  [1..] :: [Integer] should always have a tail, I would
surely like to know with an error if for some reason it stops doing so!